Move logging state into nested class
