+++ sdk/dotnet/Pulumi/Deployment/ILogger.cs
@@ -1,0 +1,16 @@
+﻿// Copyright 2016-2018, Pulumi Corporation
+
+using System.Threading.Tasks;
+
+namespace Pulumi
+{
+    internal interface ILogger
+    {
+        bool LoggedErrors { get; }
+
+        Task DebugAsync(string message, Resource? resource = null, int? streamId = null, bool? ephemeral = null);
+        Task InfoAsync(string message, Resource? resource = null, int? streamId = null, bool? ephemeral = null);
+        Task WarnAsync(string message, Resource? resource = null, int? streamId = null, bool? ephemeral = null);
+        Task ErrorAsync(string message, Resource? resource = null, int? streamId = null, bool? ephemeral = null);
+    }
+}
