+++ sdk/dotnet/Pulumi/Log.cs
@@ -13,26 +13,26 @@ namespace Pulumi
         /// Logs a debug-level message that is generally hidden from end-users.
         /// </summary>
         public static void Debug(string message, Resource? resource = null, int? streamId = null, bool? ephemeral = null)
-            => Deployment.InternalInstance.DebugAsync(message, resource, streamId, ephemeral);
+            => Deployment.InternalInstance.Logger.DebugAsync(message, resource, streamId, ephemeral);
 
         /// <summary>
         /// Logs an informational message that is generally printed to stdout during resource
         /// operations.
         /// </summary>
         public static void Info(string message, Resource? resource = null, int? streamId = null, bool? ephemeral = null)
-            => Deployment.InternalInstance.InfoAsync(message, resource, streamId, ephemeral);
+            => Deployment.InternalInstance.Logger.InfoAsync(message, resource, streamId, ephemeral);
 
         /// <summary>
         /// Warn logs a warning to indicate that something went wrong, but not catastrophically so.
         /// </summary>
         public static void Warn(string message, Resource? resource = null, int? streamId = null, bool? ephemeral = null)
-            => Deployment.InternalInstance.WarnAsync(message, resource, streamId, ephemeral);
+            => Deployment.InternalInstance.Logger.WarnAsync(message, resource, streamId, ephemeral);
 
         /// <summary>
         /// Error logs a fatal error to indicate that the tool should stop processing resource
         /// operations immediately.
         /// </summary>
         public static void Error(string message, Resource? resource = null, int? streamId = null, bool? ephemeral = null)
-            => Deployment.InternalInstance.ErrorAsync(message, resource, streamId, ephemeral);
+            => Deployment.InternalInstance.Logger.ErrorAsync(message, resource, streamId, ephemeral);
     }
 }
